person_microservice/pyproject.toml
[tool.poetry]
name = "persons_microservice"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
mysqlclient = "^2.1.0"
PyJWT = "^2.3.0"
redis = "^4.1.0"
Flask-Cors = "^3.0.10"
python-dotenv = "^0.19.2"
marshmallow = "^3.15.0"
black = "^22.3.0"
SQLAlchemy-Utils = "^0.38.2"
celery = "^5.2.6"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
